requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
pillow
pytesseract
easyocr
scikit-learn
sentence-transformers
faiss-cpu
pandas

